+++ e2e/__tests__/beforeEachQueue.ts
@@ -5,8 +5,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import {skipSuiteOnJestCircus} from '@jest/test-utils';
 import {wrap} from 'jest-snapshot-serializer-raw';
 import runJest from '../runJest';
+
+skipSuiteOnJestCircus(); // Circus does not support funky async definitions
 
 describe('Correct beforeEach order', () => {
   it('ensures the correct order for beforeEach', () => {